requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
#pkg-resources==0.0.0
python-gnupg==0.4.6
pytz==2021.1
sqlparse==0.4.1
uWSGI==2.0.17
